refactor(auth): move MCP OAuth server routes
