+++ src/treg/api.py
@@ -132,8 +132,13 @@ from .routers.auth import (
     CliApproveIn,
     EmailStartIn,
     EmailVerifyIn,
+    GrantTeamIn,
+    OAuthClientRegistration,
     _AUTH_HEAD,
+    _CONSENT_CSS,
     _auth_page,
+    _authorize_request,
+    _consent_page,
     _finish_oauth_login,
     _intercom_user_hash,
     _login_callback_base,
@@ -150,7 +155,13 @@ from .routers.auth import (
     _login_page_html,
     _norm_pair_code,
     _orgs_brief,
+    _oauth_error,
     _prune_handshakes,
+    _refresh_grant,
+    _resolve_oauth_client,
+    _same_mcp_resource,
+    _wrong_resource,
+    AUTH_CODE_TTL_S,
     auth_cli_approve,
     auth_cli_orgs,
     auth_cli_poll,
@@ -166,6 +177,16 @@ from .routers.auth import (
     auth_logout,
     auth_me,
     login_page,
+    oauth_authorization_server,
+    oauth_authorize,
+    oauth_authorize_approve,
+    oauth_grant_set_team,
+    oauth_grants,
+    oauth_protected_resource,
+    oauth_register,
+    oauth_revoke,
+    oauth_token,
+    openai_apps_challenge,
 )
 from .routers.auth_helpers import (
     OAUTH_RETURN_COOKIE,
@@ -670,588 +691,7 @@ router.routes.extend(auth_routes.invite_router.routes)
 # Register the moved site routes at their original position.
 router.routes.extend(web_routes.site_router.routes)
 
-class OAuthClientRegistration(BaseModel):
-    """RFC 7591 registration request. Extra fields are ignored rather than refused — clients send
-    plenty we do not use, and rejecting an unknown key would break them for no benefit."""
-
-    client_name: str = ""
-    redirect_uris: list[str] = []
-    client_uri: str = ""
-    logo_uri: str = ""
-    scope: str = ""
-
-
-@app.post("/oauth/register", include_in_schema=False)
-async def oauth_register(body: OAuthClientRegistration,
-                         db: AsyncSession = Depends(get_session)) -> JSONResponse:
-    """Dynamic client registration (RFC 7591) — how Claude Code and most MCP clients arrive.
-
-    Open by design: the spec has clients register unauthenticated, and a registration grants nothing
-    on its own. Every token still requires a human to sign in and approve at the consent screen, so
-    the worst a spurious registration achieves is a row in a table.
-
-    What is NOT open is the redirect URI. It is fixed here and matched exactly at authorize time,
-    because that is where authorization codes get delivered.
-    """
-    from .domain.identity import mcp_oauth
-
-    uris = [u for u in body.redirect_uris if mcp_oauth.valid_redirect_uri(u)]
-    if not uris:
-        return JSONResponse(status_code=400, content={
-            "error": "invalid_redirect_uri",
-            "error_description": ("at least one https redirect_uri is required (http is accepted "
-                                  "only for 127.0.0.1 / localhost, for CLI clients)")})
-    client = OAuthClient(
-        client_id=mcp_oauth.new_client_id(), kind="dcr",
-        client_name=(body.client_name or "unnamed client")[:200],
-        client_uri=body.client_uri[:500], logo_uri=body.logo_uri[:500],
-        redirect_uris=uris[:20], scope=body.scope[:200])
-    db.add(client)
-    await db.commit()
-    return JSONResponse(status_code=201, content={
-        "client_id": client.client_id,
-        "client_id_issued_at": int(client.created_at.timestamp()),
-        "client_name": client.client_name,
-        "redirect_uris": client.redirect_uris,
-        "grant_types": ["authorization_code", "refresh_token"],
-        "response_types": ["code"],
-        "token_endpoint_auth_method": "none",
-    })
-
-
-async def _resolve_oauth_client(client_id: str, db: AsyncSession) -> OAuthClient | None:
-    """One client row, whichever door it came through.
-
-    A registered client is a lookup. A client_id that is an https URL is a metadata document: fetched
-    on first sight, cached as a row, and refreshed when stale — documents change, and a cache that
-    never expires would pin a client to redirect URIs it has since retired.
-    """
-    from .domain.identity import mcp_oauth
-
-    row = (await db.execute(select(OAuthClient).where(OAuthClient.client_id == client_id))
-           ).scalar_one_or_none()
-    fresh_enough = row is not None and (
-        row.kind != "cimd" or (row.refreshed_at is not None and
-                               (datetime.now(timezone.utc).replace(tzinfo=None) - row.refreshed_at
-                                ).total_seconds() < mcp_oauth._CIMD_REFRESH_S))
-    if fresh_enough:
-        return row
-    if not client_id.startswith("https://"):
-        return row  # a dcr client we do not know is simply unknown
-    doc = await mcp_oauth.fetch_client_id_metadata(client_id)
-    if doc is None:
-        return row  # keep a stale copy over nothing: a transient fetch failure is not a revocation
-    now = datetime.now(timezone.utc).replace(tzinfo=None)
-    if row is None:
-        row = OAuthClient(client_id=client_id, kind="cimd")
-        db.add(row)
-    row.kind, row.refreshed_at = "cimd", now
-    row.client_name, row.client_uri = doc["client_name"], doc["client_uri"]
-    row.logo_uri, row.redirect_uris, row.scope = doc["logo_uri"], doc["redirect_uris"], doc["scope"]
-    await db.commit()
-    await db.refresh(row)
-    return row
-
-
-AUTH_CODE_TTL_S = 300   # a code is redeemed within seconds; five minutes is generous, not a window
-_CONSENT_CSS = """
-.consent{max-width:460px;text-align:left}
-.consent h1{font-size:20px;margin:0 0 4px}
-.consent .who{color:var(--ink55,#8a8a8a);font-size:13.5px;margin:0 0 18px}
-.consent .grants{list-style:none;padding:0;margin:0 0 18px}
-.consent .grants li{padding:7px 0 7px 22px;position:relative;font-size:13.5px;line-height:1.45}
-.consent .grants li:before{content:"›";position:absolute;left:6px;color:#e0703f}
-.consent label{display:block;font-size:12px;text-transform:uppercase;letter-spacing:.06em;
-  color:var(--ink55,#8a8a8a);margin:0 0 6px}
-.consent select{width:100%;padding:9px 10px;border-radius:8px;font:inherit;font-size:14px;
-  background:#1a1a1a;color:inherit;border:1px solid #333;margin-bottom:16px}
-.consent .row{display:flex;gap:10px}
-.consent button{flex:1;padding:10px 14px;border-radius:8px;font:inherit;font-size:14px;cursor:pointer;
-  border:1px solid #333;background:#1a1a1a;color:inherit}
-.consent button.primary{background:#e0703f;border-color:#e0703f;color:#161310;font-weight:600}
-.consent .fine{color:var(--ink55,#8a8a8a);font-size:12px;margin:14px 0 0;line-height:1.5}
-"""
-
-
-def _consent_page(*, client_name: str, client_uri: str, user_email: str, teams: list,
-                  hidden: dict, unverified: bool) -> HTMLResponse:
-    """The one place a human sees what they are granting — so it says it in words, not scopes.
-
-    Deliberately plain about the two things that cost money or leak data: this client will be able to
-    spend the team's balance, and to use the keys the team registered. A consent screen that lists
-    `treg:call` and calls it informed is a formality, not a decision.
-
-    The team picker is here rather than anywhere else because this is the only moment a human is
-    present to answer it. `balance` used to have to refuse and ask when someone belonged to several
-    teams; that question belongs at the grant, once.
-    """
-    import html as _h
-
-    def _label(t: dict) -> str:
-        bal = t.get("balance_usd")
-        if bal is None:
-            money = ""
-        elif bal <= 0:
-            # Named, not hidden: an empty team is still a legitimate choice when the work uses the
-            # team's OWN keys, which are never metered. Saying "no balance" is the useful warning;
-            # removing the option would be wrong.
-            money = "  ·  no balance — catalog calls will be refused"
-        else:
-            money = f"  ·  ${bal:.2f}"
-        return f'{t["slug"]} — {t["role"]}{money}'
-
-    opts = "".join(
-        f'<option value="{t["org_id"]}">{_h.escape(_label(t))}</option>' for t in teams)
-    fields = "".join(
-        f'<input type="hidden" name="{_h.escape(k)}" value="{_h.escape(str(v))}"/>'
-        for k, v in hidden.items())
-    who = _h.escape(client_name or "An application")
-    where = (f' <span class="who">({_h.escape(client_uri)})</span>' if client_uri else "")
-    warn = ("" if not unverified else
-            '<p class="fine"><b>This application registered itself.</b> treg has not reviewed it — '
-            'only continue if you recognise it and started this yourself.</p>')
-    return HTMLResponse(
-        f"{_AUTH_HEAD.replace('</style>', _CONSENT_CSS + '</style>')}"
-        f'<body><div class="wrap"><div class="card consent">'
-        f'<div class="logo">▚ tools-registry</div>'
-        f"<h1>{who} wants to use treg</h1>{where}"
-        f'<p class="who">Signed in as {_h.escape(user_email)}</p>'
-        f'<ul class="grants">'
-        f"<li>Search the catalog and read prices</li>"
-        f"<li>Call tools on your team's behalf — <b>this spends the team's balance</b></li>"
-        f"<li>Use the API keys and connections your team has registered, without seeing them</li>"
-        f"<li>Read the team's balance</li>"
-        f"</ul>"
-        f'<form method="post" action="/oauth/authorize">{fields}'
-        f'<label for="org_id">Which team?</label>'
-        f'<select id="org_id" name="org_id" required>{opts}</select>'
-        f'<div class="row">'
-        f'<button type="submit" name="decision" value="deny">Cancel</button>'
-        f'<button type="submit" name="decision" value="allow" class="primary">Allow</button>'
-        f"</div></form>{warn}"
-        f'<p class="fine">You can revoke this at any time from your treg dashboard.</p>'
-        f"</div></div></body></html>")
-
-
-def _wrong_resource(resource: str) -> str | None:
-    """Is this `resource` one we actually protect? Returns an error message, or None if fine.
-
-    Refusing early matters more than it looks. `resource` becomes the token's audience, and the MCP
-    server accepts only its own — so accepting a resource we do not serve mints a token that is
-    valid, well-formed, and silently useless. That failure surfaces later, at the first tool call,
-    as "not signed in", which points the reader at authentication when the real problem was the
-    audience. Found exactly that way: an independent MCP client sent the URL it was connecting to
-    rather than the canonical identifier from our metadata, and got a token that could never work.
-
-    Empty is allowed: a client that omits `resource` gets our canonical one, which is what it would
-    have discovered anyway.
-    """
-    from .domain.identity import mcp_oauth
-
-    if not resource:
-        return None
-    canonical = mcp_oauth.mcp_resource_url()
-    # The legacy hosts' resource URLs stay valid: a pre-move client discovered its `resource` from
-    # the old domain's metadata and will keep sending it for the lifetime of the grant.
-    if any(resource.rstrip("/") == aud.rstrip("/") for aud in mcp_oauth.mcp_resource_audiences()):
-        return None
-    return (f"this server issues tokens for {canonical} only — use the `resource` value from "
-            f"/.well-known/oauth-protected-resource")
-
-
-def _same_mcp_resource(a: str, b: str) -> bool:
-    """Whether two `resource` values name this same MCP server. Exact match, slash-variant match,
-    or BOTH normalize into the canonical+legacy audience set — the domain move renamed the
-    resource without changing it, so a grant consented on one name must stay exchangeable and
-    refreshable by a client re-based onto the other (in either direction)."""
-    from .domain.identity import mcp_oauth
-
-    na, nb = mcp_oauth.normalize_resource(a), mcp_oauth.normalize_resource(b)
-    if a == b or na == nb:
-        return True
-    auds = mcp_oauth.mcp_resource_audiences()
-    return na in auds and nb in auds
-
-
-def _oauth_error(redirect_uri: str, state: str, error: str, desc: str = ""):
-    """OAuth errors go BACK TO THE CLIENT via the redirect, once we trust the redirect.
-
-    Before the client and redirect_uri are validated we must NOT redirect — bouncing an error to an
-    unvalidated URI is an open redirect, and it would leak `state` to whoever asked for it. Those
-    cases raise a plain 400 instead, which is why this helper is only ever called after validation.
-    """
-    from urllib.parse import urlencode
-
-    q = {"error": error}
-    if desc:
-        q["error_description"] = desc
-    if state:
-        q["state"] = state
-    sep = "&" if "?" in redirect_uri else "?"
-    return RedirectResponse(f"{redirect_uri}{sep}{urlencode(q)}", status_code=302)
-
-
-async def _authorize_request(client_id: str, redirect_uri: str, response_type: str,
-                             code_challenge: str, code_challenge_method: str,
-                             db: AsyncSession):
-    """Validate an authorization request. Returns (client, error_response) — exactly one is None.
-
-    Order matters and is deliberate: identify the client and its redirect FIRST, because until both
-    are known-good there is nowhere safe to send an error. Everything after that can be reported to
-    the client properly.
-    """
-    from .domain.identity import mcp_oauth
-
-    client = await _resolve_oauth_client(client_id, db) if client_id else None
-    if client is None:
-        return None, JSONResponse(status_code=400, content={
-            "error": "invalid_client",
-            "error_description": "unknown client_id — register first, or serve a client-id metadata document"})
-    if not mcp_oauth.redirect_uri_allowed(client, redirect_uri):
-        # NOT a redirect: we do not bounce errors to a URI the client has not proven is theirs.
-        return None, JSONResponse(status_code=400, content={
-            "error": "invalid_request",
-            "error_description": "redirect_uri does not exactly match one registered for this client"})
-    return client, None
-
-
-@app.get("/oauth/authorize", include_in_schema=False)
-async def oauth_authorize(
-    request: Request,
-    client_id: str = Query(default=""), redirect_uri: str = Query(default=""),
-    response_type: str = Query(default="code"), scope: str = Query(default=""),
-    state: str = Query(default=""), code_challenge: str = Query(default=""),
-    code_challenge_method: str = Query(default=""), resource: str = Query(default=""),
-    treg_session: str = Cookie(default=""), db: AsyncSession = Depends(get_session),
-):
-    """What is this client asking for, and on behalf of which team?
-
-    Step 3 answers that as JSON; step 4 puts a consent page on top of the same checks. It issues
-    NOTHING — approval is a POST, because a GET that granted access could be triggered by any page
-    that can make the browser navigate.
-    """
-    client, err = await _authorize_request(client_id, redirect_uri, response_type,
-                                           code_challenge, code_challenge_method, db)
-    if err is not None:
-        return err
-    if response_type != "code":
-        return _oauth_error(redirect_uri, state, "unsupported_response_type",
-                            "only the authorization code flow is supported")
-    if not code_challenge or code_challenge_method != "S256":
-        return _oauth_error(redirect_uri, state, "invalid_request",
-                            "PKCE with code_challenge_method=S256 is required")
-    if (bad_target := _wrong_resource(resource)) is not None:
-        return _oauth_error(redirect_uri, state, "invalid_target", bad_target)
-
-    user = await _user_from_session(treg_session, db)
-    if user is None:
-        # Sign in first, then come back to THIS request. Parked in a cookie rather than a `?next=`
-        # query the sign-in page would have to understand — the first version invented that
-        # convention and nothing implemented it, so the user signed in and landed on the dashboard
-        # with the authorization silently dropped.
-        resp = RedirectResponse("/", status_code=302)
-        _remember_oauth_return(resp, request)
-        return resp
-
-    memberships = (await db.execute(
-        select(Membership).where(Membership.user_id == user.id))).scalars().all()
-    teams = []
-    for m in memberships:
-        org = await db.get(Org, m.org_id)
-        if org is not None and not org.suspended:
-            # The BALANCE belongs on this list. Choosing a team here decides which balance the
-            # client spends for the life of the grant, and a list of slugs makes the one question
-            # that matters — "which of these can actually pay?" — invisible at the moment of
-            # choosing. Unclecode picked a $0.00 team on the first real ChatGPT connect and the call
-            # was refused; nothing on the page could have told him.
-            teams.append({"org_id": org.id, "slug": org.slug, "role": m.role,
-                          "balance_usd": round((org.balance_micro or 0) / 1_000_000, 4)})
-    if not teams:
-        return _oauth_error(redirect_uri, state, "access_denied",
-                            "this account is not a member of any team")
-
-    hidden = {"client_id": client_id, "redirect_uri": redirect_uri, "response_type": response_type,
-              "scope": scope, "state": state, "code_challenge": code_challenge,
-              "code_challenge_method": code_challenge_method, "resource": resource}
-    if "application/json" in (request.headers.get("accept") or ""):
-        return {
-            "client": {"client_id": client.client_id, "name": client.client_name,
-                       "uri": client.client_uri, "kind": client.kind},
-            "redirect_uri": redirect_uri, "scope": scope, "resource": resource,
-            "user": user.email,
-            # The team picker. A person may belong to several, and which one this client may spend
-            # from is a decision for the human here — not something resolved per call later.
-            "teams": teams,
-            "approve_with": "POST /oauth/authorize with the same parameters plus org_id",
-        }
-    return _consent_page(client_name=client.client_name, client_uri=client.client_uri,
-                         user_email=user.email, teams=teams, hidden=hidden,
-                         unverified=(client.kind == "dcr"))
-
-
-@app.post("/oauth/authorize", include_in_schema=False)
-async def oauth_authorize_approve(
-    request: Request,
-    decision: str = Form(default="allow"),
-    client_id: str = Form(default=""), redirect_uri: str = Form(default=""),
-    response_type: str = Form(default="code"), scope: str = Form(default=""),
-    state: str = Form(default=""), code_challenge: str = Form(default=""),
-    code_challenge_method: str = Form(default=""), resource: str = Form(default=""),
-    org_id: int = Form(default=0),
-    treg_session: str = Cookie(default=""), db: AsyncSession = Depends(get_session),
-):
-    """The human decided. On approval, mint a one-time code bound to everything that made this
-    request; on anything else, tell the client no."""
-    import secrets as _s
-
-    from .domain.identity import mcp_oauth
-
-    # The consent form is the security boundary, so the submission must have come from OUR page.
-    # Without this, a page anywhere could auto-submit a form and grant itself a team's balance —
-    # the user is signed in, so their cookie would ride along. Same guard `auth_logout` uses.
-    if not _same_origin(request):
-        raise HTTPException(status_code=403, detail=(
-            "cross-origin authorization rejected — this form must be submitted from treg's own "
-            f"consent page (saw Origin: {request.headers.get('origin') or 'none'}, "
-            f"Sec-Fetch-Site: {request.headers.get('sec-fetch-site') or 'none'})"))
-
-    client, err = await _authorize_request(client_id, redirect_uri, response_type,
-                                           code_challenge, code_challenge_method, db)
-    if err is not None:
-        return err
-    if decision != "allow":
-        # Cancel is a real answer and the client is entitled to hear it, rather than hang.
-        return _oauth_error(redirect_uri, state, "access_denied", "the user declined")
-    if (bad_target := _wrong_resource(resource)) is not None:
-        return _oauth_error(redirect_uri, state, "invalid_target", bad_target)
-    if not code_challenge or code_challenge_method != "S256":
-        return _oauth_error(redirect_uri, state, "invalid_request",
-                            "PKCE with code_challenge_method=S256 is required")
-
-    user = await _user_from_session(treg_session, db)
-    if user is None:
-        return JSONResponse(status_code=401, content={"error": "access_denied",
-                                                      "error_description": "not signed in"})
-    # The chosen team must be one this user actually belongs to — the field is client-supplied.
-    membership = (await db.execute(select(Membership).where(
-        Membership.user_id == user.id, Membership.org_id == org_id))).scalar_one_or_none()
-    if membership is None:
-        return _oauth_error(redirect_uri, state, "access_denied",
-                            "choose a team you are a member of")
-
-    code = OAuthCode(
-        code=_s.token_urlsafe(32), client_id=client.client_id, user_id=user.id, org_id=org_id,
-        redirect_uri=redirect_uri, code_challenge=code_challenge,
-        resource=mcp_oauth.normalize_resource(resource) if resource else mcp_oauth.mcp_resource_url(),
-        scope=scope,
-        expires_at=datetime.now(timezone.utc).replace(tzinfo=None)
-        + timedelta(seconds=AUTH_CODE_TTL_S))
-    db.add(code)
-    await db.commit()
-
-    from urllib.parse import urlencode
-
-    q = {"code": code.code}
-    if state:
-        q["state"] = state
-    sep = "&" if "?" in redirect_uri else "?"
-    return RedirectResponse(f"{redirect_uri}{sep}{urlencode(q)}", status_code=302)
-
-
-async def _refresh_grant(*, refresh_token: str, client_id: str, resource: str,
-                         db: AsyncSession, bad):
-    """Exchange a refresh token for a new access token, ROTATING the refresh token as we go.
-
-    The retired row is kept, not deleted. That is what makes a replay recognisable: a deleted token
-    looks merely unknown, while a retired one tells us somebody used a credential that had already
-    been spent — and at that point the safe reading is that it was copied.
-    """
-    from .domain.identity import mcp_oauth
-
-    if not refresh_token:
-        return bad("invalid_request", "refresh_token is required")
-    row = (await db.execute(select(OAuthRefresh).where(
-        OAuthRefresh.token_hash == crypto.hash_token(refresh_token)))).scalar_one_or_none()
-    if row is None:
-        return bad("invalid_grant", "unknown refresh token")
-
-    if row.retired_at is not None:
-        # Already spent. Either a client retried after a dropped response, or someone else has a
-        # copy — indistinguishable from here, so assume the worse one and end the whole family. The
-        # cost of being wrong is one sign-in; the cost of the other mistake is somebody's balance.
-        killed = await _revoke_refresh_family(row.family_id, "reuse detected", db)
-        await db.commit()
-        audit.record_call(org_id=row.org_id, user_email="", tool_name="oauth.refresh_reuse",
-                          method="POST", path="/oauth/token", status_code=400, client="",
-                          telemetry={"family": row.family_id, "revoked": killed})
-        return bad("invalid_grant",
-                   "this refresh token was already used — the grant has been revoked, sign in again")
-
-    if not _refresh_is_live(row):
-        return bad("invalid_grant", "refresh token expired")
-    if client_id and client_id != row.client_id:
-        return bad("invalid_grant", "refresh token was issued to a different client")
-    if resource and not _same_mcp_resource(resource, row.resource):
-        return bad("invalid_target", "resource does not match the one that was consented to")
-
-    user = await db.get(User, row.user_id)
-    if user is None or user.suspended:
-        return bad("invalid_grant", "the account behind this grant is no longer active")
-    live_org_id = await _family_org(row.family_id, db) or row.org_id
-    org = await db.get(Org, live_org_id)
-    if org is None or org.suspended:
-        return bad("invalid_grant", "the team on this grant is no longer available")
-    # STILL a member? The grant is the user's consent to spend a TEAM's balance, and leaving (or
-    # being removed from) that team ends the standing they consented with. Without this a grant
-    # kept minting tokens forever: every downstream call was refused by `require_member`, so the
-    # damage was bounded, but the grant lay dormant and sprang back to life — with no new consent —
-    # the day the membership was restored.
-    still_in = (await db.execute(select(Membership).where(
-        Membership.user_id == row.user_id, Membership.org_id == live_org_id))).scalar_one_or_none()
-    if still_in is None:
-        await _revoke_refresh_family(row.family_id, "membership ended", db)
-        await db.commit()
-        return bad("invalid_grant",
-                   "the account behind this grant is no longer a member of its team — sign in again")
-
-    now = datetime.now(timezone.utc).replace(tzinfo=None)
-    row.retired_at, row.retired_reason = now, "rotated"
-    db.add(row)
-    replacement = await _issue_refresh(family_id=row.family_id, client_id=row.client_id,
-                                       user_id=row.user_id, org_id=live_org_id,
-                                       resource=row.resource, scope=row.scope, db=db)
-    access = mcp_oauth.make_access_token(
-        user_id=row.user_id, org_id=live_org_id, scope=row.scope,
-        audience=mcp_oauth.normalize_resource(row.resource),  # heal pre-normalization spellings
-        token_version=user.token_version)
-    await db.commit()
-    return JSONResponse({"access_token": access, "token_type": "Bearer",
-                         "expires_in": mcp_oauth.ACCESS_TTL_SECONDS, "scope": row.scope,
-                         "refresh_token": replacement})
-
-
-@app.post("/oauth/revoke", include_in_schema=False)
-async def oauth_revoke(token: str = Form(default=""),
-                       db: AsyncSession = Depends(get_session)) -> JSONResponse:
-    """RFC 7009. Revoking a refresh token ends its whole family — a user who disconnects an app
-    means all of it, not the one string they happened to send.
-
-    Always answers 200, as the RFC requires: an unknown token is already revoked as far as the caller
-    is concerned, and saying otherwise would turn this into an oracle for guessing valid tokens.
-    """
-    if token:
-        row = (await db.execute(select(OAuthRefresh).where(
-            OAuthRefresh.token_hash == crypto.hash_token(token)))).scalar_one_or_none()
-        if row is not None:
-            await _revoke_refresh_family(row.family_id, "revoked by client", db)
-            await db.commit()
-    return JSONResponse({"ok": True})
-
-
-@app.post("/oauth/token", include_in_schema=False)
-async def oauth_token(
-    grant_type: str = Form(default=""), code: str = Form(default=""),
-    redirect_uri: str = Form(default=""), client_id: str = Form(default=""),
-    code_verifier: str = Form(default=""), resource: str = Form(default=""),
-    refresh_token: str = Form(default=""),
-    db: AsyncSession = Depends(get_session),
-):
-    """Exchange a code for an access token.
-
-    Errors here are JSON, not redirects: this is a back-channel call from the client itself, and
-    there is no browser to send anywhere.
-    """
-    from .domain.identity import mcp_oauth
-
-    def bad(err: str, desc: str, status: int = 400):
-        return JSONResponse(status_code=status,
-                            content={"error": err, "error_description": desc})
-
-    if grant_type == "refresh_token":
-        return await _refresh_grant(refresh_token=refresh_token, client_id=client_id,
-                                    resource=resource, db=db, bad=bad)
-    if grant_type != "authorization_code":
-        return bad("unsupported_grant_type",
-                   "supported grants: authorization_code, refresh_token")
-
-    row = (await db.execute(select(OAuthCode).where(OAuthCode.code == code))
-           ).scalar_one_or_none() if code else None
-    if row is None:
-        return bad("invalid_grant", "unknown or already-redeemed code")
-
-    # DELETE FIRST. A code is single-use, and holding it while validating leaves a window where two
-    # redemptions both read it. Everything below is validated against values already in hand.
-    await db.delete(row)
-    await db.commit()
-
-    if row.expires_at < datetime.now(timezone.utc).replace(tzinfo=None):
-        return bad("invalid_grant", "code expired")
-    if client_id and client_id != row.client_id:
-        return bad("invalid_grant", "code was issued to a different client")
-    if redirect_uri != row.redirect_uri:
-        return bad("invalid_grant", "redirect_uri does not match the one the code was issued for")
-    if not mcp_oauth.verify_pkce(code_verifier, row.code_challenge):
-        return bad("invalid_grant", "code_verifier does not match the code_challenge")
-    if resource and not _same_mcp_resource(resource, row.resource):
-        return bad("invalid_target", "resource does not match the one that was consented to")
-
-    user = await db.get(User, row.user_id)
-    if user is None or user.suspended:
-        return bad("invalid_grant", "the account behind this grant is no longer active")
-
-    token = mcp_oauth.make_access_token(
-        user_id=row.user_id, org_id=row.org_id, scope=row.scope,
-        audience=mcp_oauth.normalize_resource(row.resource),  # heal pre-normalization spellings
-        token_version=user.token_version)
-    import secrets as _s
-
-    refresh = await _issue_refresh(family_id=_s.token_urlsafe(16), client_id=row.client_id,
-                                   user_id=row.user_id, org_id=row.org_id, resource=row.resource,
-                                   scope=row.scope, db=db)
-    await db.commit()
-    return JSONResponse({"access_token": token, "token_type": "Bearer",
-                         "expires_in": mcp_oauth.ACCESS_TTL_SECONDS, "scope": row.scope,
-                         "refresh_token": refresh})
-
-
-@app.get("/.well-known/oauth-protected-resource", include_in_schema=False)
-@app.get("/.well-known/oauth-protected-resource/mcp", include_in_schema=False)
-async def oauth_protected_resource():
-    """Tells an MCP client which authorization server guards /mcp/ and what it may ask for.
-
-    Two paths for one document: the spec has clients look it up either at the host root or under the
-    resource's own path, and which one a given client tries is not something we get to choose.
-    """
-    from .domain.identity import mcp_oauth
-
-    return JSONResponse(mcp_oauth.protected_resource_metadata(),
-                        headers={"Cache-Control": "public, max-age=3600"})
-
-
-@app.get("/.well-known/oauth-authorization-server", include_in_schema=False)
-async def oauth_authorization_server():
-    """How to get a token: the authorize and token endpoints, S256, and that we accept both dynamic
-    registration and a client-id metadata document."""
-    from .domain.identity import mcp_oauth
-
-    return JSONResponse(mcp_oauth.authorization_server_metadata(),
-                        headers={"Cache-Control": "public, max-age=3600"})
-
-
-@app.get("/.well-known/openai-apps-challenge", include_in_schema=False)
-async def openai_apps_challenge():
-    """Domain-verification token for the OpenAI plugin directory.
-
-    The portal issues a token and fetches it here to confirm we control the host serving the MCP
-    endpoint. It must return THAT token as plain text and nothing else — the documentation is
-    explicit that JSON, a list, or several tokens all fail. 404 when unset, which is correct for
-    every deployment that is not ours: an empty file would read as a verification that never
-    completes.
-    """
-    token = (get_settings().openai_apps_challenge or "").strip()
-    if not token:
-        raise HTTPException(status_code=404, detail="not configured")
-    return PlainTextResponse(token, headers={"Cache-Control": "no-store"})
+router.routes.extend(auth_routes.oauth_server_router.routes)
 
 
 # Register the moved public-document routes at their original position.
@@ -1937,99 +1377,7 @@ async def _redeem_referral(db: AsyncSession, request: Request, user: User, org: 
         logging.getLogger("treg").warning("referral attribution failed for org %s: %s", org.id, exc)
 
 
-class GrantTeamIn(BaseModel):
-    team: str = ""          # the slug (or numeric id) of a team the signed-in user belongs to
-
-
-@app.get("/oauth/grants", include_in_schema=False)
-async def oauth_grants(user: User = Depends(require_identity),
-                       db: AsyncSession = Depends(get_session)) -> list[dict]:
-    """The MCP connections this account has granted, and which team each one spends from.
-
-    The team on a grant was chosen once, on a consent screen, and after that it was invisible from
-    every side: the client reports a slug, the CLI lists the teams of whichever identity is logged
-    in THERE, and the two need not be the same account at all. Somebody spent from a team they could
-    not see listed anywhere and had no way to recognise as wrong.
-    """
-    rows = (await db.execute(select(OAuthRefresh).where(
-        OAuthRefresh.user_id == user.id, OAuthRefresh.retired_at.is_(None)
-    ).order_by(OAuthRefresh.created_at.desc()))).scalars().all()
-    rows = [row for row in rows if _refresh_is_live(row)]
-    seen: set[str] = set()
-    out: list[dict] = []
-    for row in rows:                      # one entry per GRANT, not per rotation of its token
-        if row.family_id in seen:
-            continue
-        seen.add(row.family_id)
-        # Listing and refresh must read the SAME family authority. A token row's org_id is immutable
-        # issue provenance and may legitimately name the team used before a later move.
-        grant = await _ensure_grant(row.family_id, db)
-        org_id = grant.current_org_id if grant is not None else None
-        org = await db.get(Org, org_id) if org_id is not None else None
-        client = (await db.execute(select(OAuthClient).where(
-            OAuthClient.client_id == row.client_id))).scalar_one_or_none()
-        out.append({
-            "grant": row.family_id,
-            "client": (client.client_name if client else "") or row.client_id,
-            "team": org.slug if org else None,
-            "team_name": org.name if org else None,
-            "granted": grant.granted_at.isoformat(timespec="seconds") if grant else None,
-        })
-    # GET normally reads only, but repairing a family created by an old rolling-deploy instance is
-    # a durable compatibility backfill. Without this commit the response looks healed once while
-    # the inserted authority row is rolled back when the request session closes.
-    await db.commit()
-    return out
-
-
-@app.post("/oauth/grants/{family_id}/team", include_in_schema=False)
-async def oauth_grant_set_team(family_id: str, body: GrantTeamIn,
-                               user: User = Depends(require_identity),
-                               db: AsyncSession = Depends(get_session)) -> dict:
-    """Re-point a live grant at another of the user's teams — without re-doing the OAuth dance.
-
-    The team is stored on the refresh family rather than only inside the issued access token, so
-    moving it is a row update and the next refresh picks it up. Re-consenting works too, but it
-    means disconnecting a working connector in whatever client holds it, and "the only way to fix
-    which balance this spends is to tear the connection down" is not an answer for the person who
-    just found out they were spending from the wrong one.
-
-    Only the GRANT'S OWN user may move it, and only to a team THEY are a member of — a grant must
-    never become a way to reach a team the consent screen would not have offered.
-    """
-    from .domain.identity import mcp_oauth
-
-    rows = (await db.execute(select(OAuthRefresh).where(
-        OAuthRefresh.family_id == family_id, OAuthRefresh.user_id == user.id,
-        OAuthRefresh.retired_at.is_(None)))).scalars().all()
-    rows = [row for row in rows if _refresh_is_live(row)]
-    if not rows:
-        raise HTTPException(status_code=404, detail=f"no live grant {family_id!r} on this account")
-    org = await _resolve_org(body.team, db)
-    member = (await db.execute(select(Membership).where(
-        Membership.user_id == user.id, Membership.org_id == org.id))).scalar_one_or_none() if org else None
-    # ONE answer for "no such team" and "a team that isn't yours". Told apart, this route reports
-    # whether an arbitrary slug exists on treg — a slug-existence oracle any signed-in account could
-    # walk. The caller's own teams are already listed to them by `treg org ls`, so the distinction
-    # buys them nothing they cannot see elsewhere.
-    if org is None or org.suspended or member is None:
-        raise HTTPException(status_code=404, detail=(
-            f"no team {body.team!r} on this account — see `treg org ls` for the teams you can use"))
-    # Change only family authority. Token rows are evidence of where each historical bearer was
-    # issued and must stay immutable, especially retired rows kept for reuse detection.
-    grant = await _ensure_grant(family_id, db)
-    if grant is None:  # the live-row check above makes this defensive, not a normal outcome
-        raise HTTPException(status_code=404, detail=f"no live grant {family_id!r} on this account")
-    grant.current_org_id = org.id
-    db.add(grant)
-    await db.commit()
-    return {
-        "grant": family_id, "team": org.slug, "team_name": org.name,
-        # Access tokens live an hour and carry the old team until the client refreshes. Saying so
-        # is the difference between "it didn't work" and "it hasn't taken effect yet".
-        "note": (f"new calls spend from {org.slug!r} once the client refreshes its access token "
-                 f"(within {mcp_oauth.ACCESS_TTL_SECONDS // 60} minutes)"),
-    }
+router.routes.extend(auth_routes.grants_router.routes)
 
 
 @app.get("/orgs")
